requirements.txt
matplotlib==3.2.0
numpy==1.18.1
scikit-learn==0.23.0
scipy==1.4.1
sklearn==0.0
torch==1.4.0
torchvision==0.5.0
Pillow==7.1.1
